# src/accessi-module/openapi/accessi.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accessi/permission/roles": {
      "get": {
        "description": "Recupera tutti i ruoli presenti nel sistema con le relative voci di menu.",
        "operationId": "getRoles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Elenco dei ruoli con i rispettivi menu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRolesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ritorna i ruoli disponibili con i relativi menu",
        "tags": [
          "Permission"
        ]
      }
    },
    "/api/accessi/permission/update-role/{codiceRuolo}": {
      "put": {
        "operationId": "updateRole",
        "parameters": [
          {
            "name": "codiceRuolo",
            "required": true,
            "in": "path",
            "description": "Codice identificativo del ruolo da aggiornare",
            "schema": {
              "example": 101,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Dati aggiornati del ruolo",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Il ruolo e stato aggiornato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore di validazione nei dati inviati",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aggiorna un ruolo esistente",
        "tags": [
          "Permission"
        ]
      }
    },
    "/api/accessi/permission/create-role": {
      "post": {
        "operationId": "createRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Dati del nuovo ruolo",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Il ruolo e stato creato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore di validazione nei dati inviati",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea un nuovo ruolo",
        "tags": [
          "Permission"
        ]
      }
    },
    "/api/accessi/permission/assign-roles/{codiceUtente}": {
      "post": {
        "operationId": "assignRolesToUser",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "description": "Codice identificativo dell'utente a cui assegnare i ruoli",
            "schema": {
              "example": 22,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Lista dei ruoli da assegnare all'utente",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolesToUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ruoli assegnati con successo all'utente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore di validazione nei dati inviati",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assegna piu ruoli a un utente",
        "tags": [
          "Permission"
        ]
      }
    },
    "/api/accessi/permission/assign-permissions/{codiceUtente}": {
      "post": {
        "operationId": "assignPermissionsToUser",
        "parameters": [
          {
            "name": "codiceUtente",
            "required": true,
            "in": "path",
            "description": "Codice identificativo dell'utente a cui assegnare le abilitazioni",
            "schema": {
              "example": 22,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Lista delle abilitazioni da assegnare all'utente",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionsToUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Abilitazioni assegnate con successo all'utente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore di validazione nei dati inviati",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assegna abilitazioni dirette a un utente",
        "tags": [
          "Permission"
        ]
      }
    },
    "/api/accessi/permission/delete-role/{codiceRuolo}": {
      "delete": {
        "operationId": "deleteRole",
        "parameters": [
          {
            "name": "codiceRuolo",
            "required": true,
            "in": "path",
            "description": "Codice identificativo del ruolo da eliminare",
            "schema": {
              "example": 382,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ruolo eliminato con successo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Errore nei parametri della richiesta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Errore interno del server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Elimina un ruolo esistente",
        "tags": [
          "Permission"
        ]
      }
    },
...
  },
  "info": {
    "title": "Accessi API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "severity": {
            "type": "string",
            "example": "error"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "statusCode": {
            "type": "number",
            "example": 2
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          },
          "code": {
            "type": "string",
            "example": "ACCESSI_DATABASE_SCHEMA_OUTDATED",
            "description": "Codice stabile leggibile dal client. Non contiene dettagli tecnici o dati sensibili."
          },
          "error": {
            "type": "string",
            "example": "ACCESSI_REQUEST_ERROR",
            "description": "Campo legacy che riporta il codice stabile dell errore, senza SQL o stack trace."
          },
          "details": {
            "example": [
              "email: deve essere una stringa."
            ],
            "description": "Dettagli di validazione sicuri, presenti solo per ACCESSI_VALIDATION_ERROR.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "severity",
          "status",
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "GetRolesResponse": {
        "type": "object",
        "properties": {
          "Status": {
            "$ref": "#/components/schemas/Status"
          },
          "Message": {
            "type": "string",
            "example": "Dati recuperati con successo.",
            "description": "Messaggio informativo restituito dalla libreria quando presente."
          },
          "Result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        },
        "required": [
          "Status",
          "Result"
        ]
      },
...
    }
  }
}